pyproject.toml
[project]
name = "mcps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.6.0",
    "fastmcp>=2.12.3",
    "psycopg2-binary>=2.9.10",
]
